pyproject.toml
[project]
name = "swarmboard"
version = "0.1.0"
description = "SwarmBoard — ZMQ multi-instance blackboard for multi-AI CLI collaboration"
requires-python = ">=3.10"
dependencies = [
    "pyzmq>=26.0",
    "mcp>=1.0.0",
    "loguru>=0.7.0",
]

[project.scripts]
swarmboard-server = "swarmboard.server:main"
swarmboard-client = "swarmboard.client:main"
swarmboard-commander = "swarmboard.commander:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/swarmboard"]
